refactor(was): 코드 간결화
