+++ src/main/java/controller/StaticController.java
@@ -3,15 +3,15 @@ package controller;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import utils.FileIoUtils;
-import utils.MyHeaders;
-import utils.MyParams;
+import model.dto.MyHeaders;
+import model.dto.MyParams;
 
 import java.io.DataOutputStream;
 import java.io.IOException;
 import java.net.URISyntaxException;
 
-import static utils.ResponseBodies.responseBody;
-import static utils.ResponseHeaders.response200Header;
+import static model.dto.ResponseBodies.responseBody;
+import static model.dto.ResponseHeaders.response200Header;
 
 public class StaticController implements MyController{
