+++ src/main/java/model/dto/MyHeaders.java
@@ -1,4 +1,4 @@
-package utils;
+package model.dto;
 
 import lombok.Getter;
