+++ src/main/java/model/dto/ResponseHeaders.java
@@ -1,4 +1,4 @@
-package utils;
+package model.dto;
 
 import org.slf4j.LoggerFactory;
